Moved VoltFunctions to the common place
